Address PR review items
