+++ components/aws/aws.app.js
@@ -1,5 +1,8 @@
 const awsRegions = require("./regions");
-const { generateRandomUniqueName } = require("./sources/common/utils");
+const {
+  generateRandomUniqueName,
+  toSingleLineString,
+} = require("./sources/common/utils");
 
 module.exports = {
   type: "app",
@@ -14,9 +17,125 @@ module.exports = {
         if (page !== 0) {
           return [];
         }
-
         return this._getAvailableRegions();
       },
+    },
+    bucket: {
+      type: "string",
+      label: "Bucket",
+      description: "The S3 bucket to watch for events",
+      async options(context) {
+        const { page = 0 } = context;
+        if (page !== 0) {
+          return [];
+        }
+        const buckets = await this._getBuckets();
+        return buckets.map((bucket) => bucket.Name);
+      },
+    },
+    secret: {
+      type: "string",
+      secret: true,
+      label: "Secret",
+      optional: true,
+      description: toSingleLineString(`
+        **Optional but recommended**: if you enter a secret here,
+        you must pass this value in [the
+        \`secret\` parameter of each HTTP POST request](https://git.io/JsJ6m)
+      `),
+    },
+    ignoreDeleteMarkers: {
+      type: "boolean",
+      label: "Ignore Delete Markers",
+      description: toSingleLineString(`
+        When ignoring delete markers this will only emit events for permanently deleted files`),
+    },
+    domain: {
+      label: "SES Domain",
+      description: "The domain you'd like to configure a catch-all handler for",
+      type: "string",
+      options() {
+        return this.sesIdentities();
+      },
+    },
+    logGroupNames: {
+      label: "CloudWatch Log Groups",
+      description: "The log groups you'd like to query",
+      type: "string[]",
+      async options({ prevContext }) {
+        const prevToken = prevContext.nextToken;
+        const {
+          logGroups,
+          nextToken,
+        } = await this.aws.logsInsightsDescibeLogGroups(this.region, prevToken);
+        const options = logGroups.map((group) => {
+          return {
+            label: group.logGroupName,
+            value: group.logGroupName,
+          };
+        });
+        return {
+          options,
+          context: {
+            nextToken,
+          },
+        };
+      },
+    },
+    queryString: {
+      label: "Logs Insights Query",
+      description: toSingleLineString(`
+        The query you'd like to run. See [this AWS
+        doc](https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/CWL_QuerySyntax.html)
+        for help with query syntax
+      `),
+      type: "string",
+    },
+    emitResultsInBatch: {
+      type: "boolean",
+      label: "Emit query results as a single event",
+      description: toSingleLineString(`
+        If \`true\`, all events are emitted as an array, within a single Pipedream event.
+        If \`false\`, each row of results is emitted as its own event.
+      `),
+    },
+    eventTypes: {
+      type: "string[]",
+      label: "Event Types",
+      description: "The type of events to watch",
+      options: [
+        // See the AWS docs for more information about the supported events
+        // emitted by S3: https://amzn.to/3AtmKy1
+        "s3:ObjectCreated:Put",
+        "s3:ObjectCreated:Post",
+        "s3:ObjectCreated:Copy",
+        "s3:ObjectCreated:CompleteMultipartUpload",
+        "s3:ObjectRemoved:Delete",
+        "s3:ObjectRemoved:DeleteMarkerCreated",
+        "s3:ObjectRestore:Post",
+        "s3:ObjectRestore:Completed",
+        "s3:ReducedRedundancyLostObject",
+        "s3:Replication:OperationFailedReplication",
+        "s3:Replication:OperationMissedThreshold",
+        "s3:Replication:OperationReplicatedAfterThreshold",
+        "s3:Replication:OperationNotTracked",
+      ],
+    },
+    topic: {
+      label: "SNS Topic Name",
+      description: toSingleLineString(`
+        **Pipedream will create an SNS topic with this name in your account**,
+        converting it to a [valid SNS topic
+        name](https://docs.aws.amazon.com/sns/latest/api/API_CreateTopic.html).
+      `),
+      type: "string",
+    },
+    detectRestoreInitiation: {
+      type: "boolean",
+      label: "Detect Restore Initiation",
+      description: toSingleLineString(`
+        When enabled, this event source will also emit events whenever a restore is initiated
+      `),
     },
   },
   methods: {
